move some tests around
