+++ .github/workflows/build.yml
@@ -44,5 +44,6 @@ jobs:
           name: test-results
           path: |
             **/target/surefire-reports/*.txt
+            **/target/surefire-reports/*.xml
             target/hyper/*.log
           retention-days: 5
