+++ jdbc-core/src/test/java/com/salesforce/datacloud/jdbc/core/partial/DataCloudQueryPollingMockTests.java
@@ -17,19 +17,26 @@ package com.salesforce.datacloud.jdbc.core.partial;
 
 import com.salesforce.datacloud.jdbc.core.DataCloudStatement;
 import com.salesforce.datacloud.jdbc.core.HyperGrpcTestBase;
+import com.salesforce.datacloud.jdbc.hyper.HyperServerConfig;
 import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.assertj.core.api.AssertionsForClassTypes;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import salesforce.cdp.hyperdb.v1.HyperServiceGrpc;
 
 import java.time.Duration;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
+import static org.grpcmock.GrpcMock.atLeast;
 import static org.grpcmock.GrpcMock.calledMethod;
 import static org.grpcmock.GrpcMock.times;
 import static org.grpcmock.GrpcMock.verifyThat;
 
+@Slf4j
 public class DataCloudQueryPollingMockTests extends HyperGrpcTestBase {
     @Test
     @SneakyThrows
@@ -45,6 +52,34 @@ public class DataCloudQueryPollingMockTests extends HyperGrpcTestBase {
                     () -> connection.waitForResultsProduced(statement.getQueryId(), Duration.ofSeconds(30)));
 
             verifyThat(calledMethod(HyperServiceGrpc.getGetQueryInfoMethod()), times(1));
+        }
+    }
+
+    @SneakyThrows
+    @ParameterizedTest
+    @ValueSource(
+            strings = {
+                    //                "SELECT PG_SLEEP(min(max(RANDOM(),5),5));",
+                    "select cast(a as numeric(38,18)) a, cast(a as numeric(38,18)) b, cast(a as numeric(38,18)) c from generate_series(1, 1024 * 1024 * 10) as s(a) order by a asc"
+            })
+    void getQueryInfoRetriesOnTimeout(String query) {
+        val configWithSleep =
+                HyperServerConfig.builder().grpcRequestTimeoutSeconds("2s").build();
+        try (val connection = getInterceptedClientConnection(configWithSleep)) {
+            val statement = connection.createStatement().unwrap(DataCloudStatement.class);
+
+            statement.execute(query);
+            val queryId = statement.getQueryId();
+
+            verifyThat(calledMethod(HyperServiceGrpc.getGetQueryInfoMethod()), times(0));
+
+            log.warn("waiting for results produced, queryId={}", queryId);
+
+            val status = connection.waitForResultsProduced(queryId, Duration.ofSeconds(60));
+
+            verifyThat(calledMethod(HyperServiceGrpc.getGetQueryInfoMethod()), atLeast(2));
+
+            assertThat(status.getQueryId()).isEqualTo(queryId);
         }
     }
 }
